backend/pyproject.toml
[project]
name = "mcp-platform-backend"
version = "0.1.0"
description = "FastAPI control-plane API for MCP Platform"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.0",
  "mangum==0.19.0",
  "boto3==1.35.36",
  "pydantic==2.9.2",
  "python-ulid==3.0.0",
  "pyjwt[crypto]==2.9.0",
  "httpx==0.27.2",
  "aws-lambda-powertools==3.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.3",
  "pytest-asyncio==0.24.0",
  "ruff==0.6.9",
  "mypy==1.11.2",
  "moto[dynamodb,secretsmanager]==5.0.16",
  "respx==0.21.1",
]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S", "W"]
# S101 = assert allowed; B008 = FastAPI Depends() in arg defaults.
ignore = ["S101", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]
# Pydantic models mirror the camelCase JSON API surface.
"app/models.py" = ["N815"]
"app/routers/*" = ["N803"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
